rename: 컴포넌트명 수정 (Property 제외)
