+++ src/app/property/[id]/components/ImageCarousel.tsx
