+++ src/app/property/[id]/components/detail/InfoBox.tsx
