+++ src/app/property/[id]/page.tsx
@@ -2,9 +2,9 @@
 
 import { useState } from "react";
 import { Tab } from "@/components/Tab";
-import PropertyDetailTab from "./components/detail/PropertyDetailTab";
-import PropertyReviewTab from "./components/review/PropertyReviewTab";
-import PropertyImageCarousel from "./components/PropertyImageCarousel";
+import DetailTab from "./components/detail/DetailTab";
+import ReviewTab from "./components/review/ReviewTab";
+import ImageCarousel from "./components/ImageCarousel";
 
 const TAB_OPTIONS = [
   { label: "상세 정보", value: "detail" },
@@ -17,15 +17,15 @@ export default function PropertyPage({ params }: { params: { id: string } }) {
   return (
     <>
       {/* 이미지 캐러셀은 상세 정보 탭에서만 보여줌 */}
-      {selectedTab === "detail" && <PropertyImageCarousel />}
+      {selectedTab === "detail" && <ImageCarousel />}
 
       {/* 탭 고정 */}
       <Tab tabOptions={TAB_OPTIONS} selected={selectedTab} onChange={setSelectedTab} />
 
       {/* 탭 콘텐츠 */}
       <div>
-        {selectedTab === "detail" && <PropertyDetailTab id={params.id} />}
-        {selectedTab === "review" && <PropertyReviewTab id={params.id} />}
+        {selectedTab === "detail" && <DetailTab id={params.id} />}
+        {selectedTab === "review" && <ReviewTab id={params.id} />}
       </div>
     </>
   );
